casos grandes d

+++ printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: agonzale <agonzale@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/23 14:01:56 by agonzale          #+#    #+#             */
-/*   Updated: 2020/10/23 17:52:26 by agonzale         ###   ########.fr       */
+/*   Updated: 2020/10/23 18:05:09 by agonzale         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -42,7 +42,7 @@ typedef struct  s_princp {
 void	ft_initialize(t_list *l);
 void	ft_specifier(t_list *l, const char *line);
 int		ft_isspecifier(char c);
-void	ft_putnbr(t_list *l, int number);
+void	ft_putnbr(t_list *l, long int number);
 void	ft_putnbr_u(t_list *l, unsigned int number);
 int     ft_strlen (t_list *l, char *str); //borrar t_list ya q no lo uso
 int		ft_get_size_num(t_list *l, int number);
